requirements.txt
pytest>=3.6.4
pip-upgrader>=1.4.15
webdriver-manager>=3.8.5
selenium>=4.8.2
pytest-html>=1.19.0
pytest-xdist>=1.22.5
pytest-cov==4.0.0
python-coveralls>=2.9.1
hypothesis>=3.78.0
assertpy>=1.1
faker>=15.3.4
allure-pytest>=2.12.0
PyYAML>=6.0
python-dotenv>=1.0.0
requests>=2.18.4
assertpy>=1.1